Reallocate styles in appropriate files
Reference https://github.com/JetBrains/jewel/pull/713
