+++ platform/jewel/ide-laf-bridge/src/main/kotlin/org/jetbrains/jewel/bridge/theme/IntUiBridgeHorizontalProgressBar.kt
@@ -1,24 +1,12 @@
 package org.jetbrains.jewel.bridge.theme
 
 import androidx.compose.foundation.shape.CornerSize
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.takeOrElse
 import androidx.compose.ui.unit.dp
 import kotlin.time.Duration.Companion.milliseconds
 import org.jetbrains.jewel.bridge.retrieveColorOrUnspecified
-import org.jetbrains.jewel.ui.component.styling.CircularProgressStyle
 import org.jetbrains.jewel.ui.component.styling.HorizontalProgressBarColors
 import org.jetbrains.jewel.ui.component.styling.HorizontalProgressBarMetrics
 import org.jetbrains.jewel.ui.component.styling.HorizontalProgressBarStyle
-
-internal fun readCircularProgressStyle(isDark: Boolean) =
-    CircularProgressStyle(
-        frameTime = 125.milliseconds,
-        color =
-            retrieveColorOrUnspecified("ProgressIcon.color").takeOrElse {
-                if (isDark) Color(0xFF6F737A) else Color(0xFFA8ADBD)
-            },
-    )
 
 internal fun readHorizontalProgressBarStyle() =
     HorizontalProgressBarStyle(
